Reorder numbers signs
The `-` in character classes may be interpreted as separator easily
(e.g.: ([0-1]`). Therefore they should be placed to the very beginning
if they are meant to match as character for sure.

It doesn't make a difference in this use case, but should probably be
applied as general rule of thumb.